pyproject.toml
[tool.poetry]
name = "capstone-project"
version = "0.1.0"
description = ""
authors = ["yash-solanki09442 <yashsolanki09442@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = ">=5.2,<6.0"
djangorestframework = "^3.16.1"
psycopg = {extras = ["binary"], version = "^3.3.3"}
python-dotenv = "^1.2.1"
psycopg2-binary = "^2.9.11"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.5.1"}

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88
